inline trips&metrics processing with rt events
