+++ python_src/src/lamp_py/performance_manager/l1_rt_metrics.py
@@ -1,443 +1,19 @@
-from typing import Dict, List, Any
-import logging
+from typing import List
 
 import numpy
+import pandas
 import sqlalchemy as sa
 
 from lamp_py.postgres.postgres_utils import DatabaseManager
 from lamp_py.postgres.postgres_schema import (
-    VehicleEvents,
-    VehicleTrips,
     VehicleEventMetrics,
 )
 from lamp_py.runtime_utils.process_logger import ProcessLogger
-from .l1_cte_statements import (
-    get_rt_trips_cte,
-    get_static_trips_cte,
-    get_trips_for_metrics,
-)
-
-
-def get_seed_data(db_manager: DatabaseManager) -> Dict[int, List[int]]:
-    """
-    get seed data for populating trip and metrics tables
-
-    distinct combinations of start_date and fk_static_timestamp allow for the pulling
-    of all records associated with a full service date
-
-    return: Dict[start_date:int : List[fk_static_timestamps]]
-    """
-    max_days_to_process = 3
-
-    last_metric_update = sa.select(
-        sa.func.coalesce(
-            sa.func.max(VehicleTrips.updated_on),
-            sa.func.to_timestamp(0),
-        )
-    ).scalar_subquery()
-
-    seed_query = (
-        sa.select(
-            VehicleEvents.start_date,
-            VehicleEvents.fk_static_timestamp,
-        )
-        .distinct(
-            VehicleEvents.start_date,
-            VehicleEvents.fk_static_timestamp,
-        )
-        .where(VehicleEvents.updated_on > last_metric_update)
-        .order_by(VehicleEvents.start_date)
-        .limit(100)
-    )
-
-    # returns empty list if no records to process
-    seed_data = db_manager.select_as_list(seed_query)
-
-    seed_data_return: Dict[Any, List[Any]] = {}
-    for record in seed_data:
-        start_date = record.get("start_date")
-        timestamp = record.get("fk_static_timestamp")
-
-        if start_date not in seed_data_return:
-            if len(seed_data_return) < max_days_to_process:
-                seed_data_return[start_date] = []
-            else:
-                break
-        seed_data_return[start_date].append(timestamp)
-
-    return seed_data_return
-
-
-def process_tables(
-    db_manager: DatabaseManager, seed_data: Dict[int, List[int]]
-) -> None:
-    """
-    process updates to multiple days of trips / metrics
-
-    during normal running this will coninually process one day at a time, but
-    during initial loading this number could be increased to speed up loading of
-    trips and metrics table
-    """
-    process_logger = ProcessLogger(
-        "l1_tables_loader", seed_data_size=len(seed_data)
-    )
-    process_logger.log_start()
-
-    days_processed = 0
-    for seed_start_date in sorted(seed_data.keys()):
-        seed_timestamps = seed_data[seed_start_date]
-        process_trips_table(db_manager, seed_start_date, seed_timestamps)
-        process_metrics_table(db_manager, seed_start_date, seed_timestamps)
-        days_processed += 1
-
-    process_logger.add_metadata(days_processed=days_processed)
-    process_logger.log_complete()
+from .l1_cte_statements import get_trips_for_metrics
 
 
 # pylint: disable=R0914
 # pylint too many local variables (more than 15)
-def process_trips_table(
-    db_manager: DatabaseManager,
-    seed_start_date: int,
-    seed_timestamps: List[int],
-) -> None:
-    """
-    processs updates to trips table
-
-    TODO: Move to L0 processing when events table no longer has trip data
-    """
-    process_logger = ProcessLogger("l1_rt_trips_table_loader")
-    process_logger.log_start()
-
-    static_trips_cte = get_static_trips_cte(seed_timestamps, seed_start_date)
-
-    # to build a 'summary' trips table only the first and last records for each
-    # static trip are needed.
-    first_stop_static_cte = (
-        sa.select(
-            static_trips_cte.c.timestamp,
-            static_trips_cte.c.static_trip_id,
-            static_trips_cte.c.static_stop_timestamp.label("static_start_time"),
-            static_trips_cte.c.parent_station.label("static_trip_first_stop"),
-        )
-        .select_from(static_trips_cte)
-        .where(static_trips_cte.c.static_trip_first_stop == sa.true())
-        .cte(name="first_stop_static_cte")
-    )
-
-    # join first_stop_static_cte with last stop records to create trip summary records
-    static_trips_summary_cte = (
-        sa.select(
-            static_trips_cte.c.timestamp,
-            static_trips_cte.c.static_trip_id,
-            static_trips_cte.c.route_id,
-            static_trips_cte.c.direction_id,
-            first_stop_static_cte.c.static_start_time,
-            first_stop_static_cte.c.static_trip_first_stop,
-            static_trips_cte.c.parent_station.label("static_trip_last_stop"),
-            static_trips_cte.c.static_trip_stop_rank.label("static_stop_count"),
-        )
-        .select_from(static_trips_cte)
-        .join(
-            first_stop_static_cte,
-            sa.and_(
-                static_trips_cte.c.timestamp
-                == first_stop_static_cte.c.timestamp,
-                static_trips_cte.c.static_trip_id
-                == first_stop_static_cte.c.static_trip_id,
-            ),
-        )
-        .where(static_trips_cte.c.static_trip_last_stop == sa.true())
-        .cte(name="static_trips_summary_cte")
-    )
-
-    rt_trips_cte = get_rt_trips_cte(seed_start_date)
-
-    # same methodology as static stop trips summary, but for RT trips
-    # pulling values associated with first stop of trips based on trip_hash
-    first_stop_rt_cte = (
-        sa.select(
-            rt_trips_cte.c.trip_hash,
-            rt_trips_cte.c.parent_station.label("rt_trip_first_stop"),
-        )
-        .select_from(rt_trips_cte)
-        .where(rt_trips_cte.c.rt_trip_first_stop_flag == sa.true())
-    ).cte(name="first_stop_rt_cte")
-
-    # join first_stop_rt_cte with values from last stop
-    rt_trips_summary_cte = (
-        sa.select(
-            rt_trips_cte.c.fk_static_timestamp.label("timestamp"),
-            rt_trips_cte.c.trip_hash,
-            rt_trips_cte.c.direction_id,
-            rt_trips_cte.c.route_id,
-            rt_trips_cte.c.start_date,
-            rt_trips_cte.c.start_time,
-            rt_trips_cte.c.vehicle_id,
-            first_stop_rt_cte.c.rt_trip_first_stop,
-            rt_trips_cte.c.parent_station.label("rt_trip_last_stop"),
-            rt_trips_cte.c.rt_trip_stop_rank.label("rt_stop_count"),
-        )
-        .select_from(rt_trips_cte)
-        .join(
-            first_stop_rt_cte,
-            rt_trips_cte.c.trip_hash == first_stop_rt_cte.c.trip_hash,
-        )
-        .where(rt_trips_cte.c.rt_trip_last_stop_flag == sa.true())
-    ).cte(name="rt_trips_summary_cte")
-
-    # perform matching between static trips summary records and RT trips summary records
-    #
-    # first match is "exact" indicating that the first stop station and last stop station
-    # are equivalent and trip start times are within one of
-    exact_trips_match = (
-        sa.select(
-            rt_trips_summary_cte.c.timestamp,
-            rt_trips_summary_cte.c.trip_hash,
-            rt_trips_summary_cte.c.direction_id,
-            rt_trips_summary_cte.c.route_id,
-            rt_trips_summary_cte.c.start_date,
-            rt_trips_summary_cte.c.start_time,
-            rt_trips_summary_cte.c.vehicle_id,
-            rt_trips_summary_cte.c.rt_stop_count,
-            static_trips_summary_cte.c.static_trip_id,
-            static_trips_summary_cte.c.static_start_time,
-            static_trips_summary_cte.c.static_stop_count,
-            sa.literal(True).label("first_last_station_match"),
-        )
-        .distinct(
-            rt_trips_summary_cte.c.trip_hash,
-        )
-        .select_from(rt_trips_summary_cte)
-        .join(
-            static_trips_summary_cte,
-            sa.and_(
-                rt_trips_summary_cte.c.timestamp
-                == static_trips_summary_cte.c.timestamp,
-                rt_trips_summary_cte.c.direction_id
-                == static_trips_summary_cte.c.direction_id,
-                rt_trips_summary_cte.c.route_id
-                == static_trips_summary_cte.c.route_id,
-                rt_trips_summary_cte.c.rt_trip_first_stop
-                == static_trips_summary_cte.c.static_trip_first_stop,
-                rt_trips_summary_cte.c.rt_trip_last_stop
-                == static_trips_summary_cte.c.static_trip_last_stop,
-                sa.func.abs(
-                    rt_trips_summary_cte.c.start_time
-                    - static_trips_summary_cte.c.static_start_time
-                )
-                < 3600,
-            ),
-            isouter=True,
-        )
-        .order_by(
-            rt_trips_summary_cte.c.trip_hash,
-            sa.func.abs(
-                rt_trips_summary_cte.c.start_time
-                - static_trips_summary_cte.c.static_start_time
-            ),
-        )
-    ).cte(name="exact_trips_match")
-
-    # backup matching logic, should match all remaining RT trips to static trips,
-    # assuming that the route_id exists in the static schedule data
-    backup_trips_match = (
-        sa.select(
-            exact_trips_match.c.timestamp,
-            exact_trips_match.c.trip_hash,
-            exact_trips_match.c.direction_id,
-            exact_trips_match.c.route_id,
-            exact_trips_match.c.start_date,
-            exact_trips_match.c.start_time,
-            exact_trips_match.c.vehicle_id,
-            exact_trips_match.c.rt_stop_count,
-            static_trips_summary_cte.c.static_trip_id,
-            static_trips_summary_cte.c.static_start_time,
-            static_trips_summary_cte.c.static_stop_count,
-            sa.literal(False).label("first_last_station_match"),
-        )
-        .distinct(
-            exact_trips_match.c.trip_hash,
-        )
-        .select_from(exact_trips_match)
-        .join(
-            static_trips_summary_cte,
-            sa.and_(
-                exact_trips_match.c.timestamp
-                == static_trips_summary_cte.c.timestamp,
-                exact_trips_match.c.direction_id
-                == static_trips_summary_cte.c.direction_id,
-                exact_trips_match.c.route_id
-                == static_trips_summary_cte.c.route_id,
-            ),
-            isouter=True,
-        )
-        .where(exact_trips_match.c.static_trip_id.is_(None))
-        .order_by(
-            exact_trips_match.c.trip_hash,
-            sa.func.abs(
-                exact_trips_match.c.start_time
-                - static_trips_summary_cte.c.static_start_time
-            ),
-        )
-    ).cte(name="backup_trips_match")
-
-    # union exact and backup trip matching records for final "new" trips summary records
-    all_new_trips = (
-        sa.union(
-            sa.select(exact_trips_match).where(
-                exact_trips_match.c.static_trip_id.is_not(None)
-            ),
-            sa.select(backup_trips_match),
-        )
-    ).cte(name="all_new_trips")
-
-    # compare all_new_trips records to existing records in RDS based on trip_hash
-    #
-    # if trip_hash does not exist in RDS, insert records
-    # if trip_hash does exist and specified fields are not equivlanet, update records
-    # update based on:
-    # - stop_count
-    # - static_trip_id_guess
-    # - static_start_time
-    # - static_stop_count
-    # - first_last_station_match (exact matching)
-    update_insert_query = (
-        sa.select(
-            all_new_trips.c.trip_hash,
-            VehicleTrips.trip_hash.label("existing_trip_hash"),
-            all_new_trips.c.direction_id,
-            all_new_trips.c.route_id,
-            sa.case(
-                [
-                    (
-                        all_new_trips.c.route_id.like("Green%"),
-                        sa.literal("Green"),
-                    ),
-                ],
-                else_=all_new_trips.c.route_id,
-            ).label("trunk_route_id"),
-            all_new_trips.c.start_date,
-            all_new_trips.c.start_time,
-            all_new_trips.c.vehicle_id,
-            all_new_trips.c.rt_stop_count.label("stop_count"),
-            all_new_trips.c.static_trip_id.label("static_trip_id_guess"),
-            all_new_trips.c.static_start_time,
-            all_new_trips.c.static_stop_count,
-            all_new_trips.c.first_last_station_match,
-            sa.case(
-                [
-                    (
-                        sa.or_(
-                            all_new_trips.c.rt_stop_count
-                            != VehicleTrips.stop_count,
-                            all_new_trips.c.static_trip_id
-                            != VehicleTrips.static_trip_id_guess,
-                            all_new_trips.c.static_start_time
-                            != VehicleTrips.static_start_time,
-                            all_new_trips.c.static_stop_count
-                            != VehicleTrips.static_stop_count,
-                            all_new_trips.c.first_last_station_match
-                            != VehicleTrips.first_last_station_match,
-                        ),
-                        sa.literal(True),
-                    )
-                ],
-                else_=sa.literal(False),
-            ).label("to_update"),
-        )
-        .select_from(all_new_trips)
-        .join(
-            VehicleTrips,
-            sa.and_(
-                all_new_trips.c.trip_hash == VehicleTrips.trip_hash,
-                VehicleTrips.start_date == seed_start_date,
-            ),
-            isouter=True,
-        )
-    )
-
-    # perform update/insert opertions with dataframe pulled from RDS
-    update_insert_trips = db_manager.select_as_dataframe(update_insert_query)
-
-    if update_insert_trips.shape[0] == 0:
-        process_logger.add_metadata(
-            trip_insert_count=0,
-            trip_update_count=0,
-        )
-        process_logger.log_complete()
-        return
-
-    update_insert_trips["start_date"] = update_insert_trips[
-        "start_date"
-    ].astype("int64")
-    update_insert_trips["start_time"] = update_insert_trips[
-        "start_time"
-    ].astype("int64")
-    update_insert_trips["stop_count"] = update_insert_trips[
-        "stop_count"
-    ].astype("int64")
-    update_insert_trips["static_start_time"] = update_insert_trips[
-        "static_start_time"
-    ].astype("Int64")
-    update_insert_trips["static_stop_count"] = update_insert_trips[
-        "static_stop_count"
-    ].astype("Int64")
-    update_insert_trips = update_insert_trips.fillna(numpy.nan).replace(
-        [numpy.nan], [None]
-    )
-
-    update_mask = (
-        update_insert_trips["existing_trip_hash"].notna()
-    ) & numpy.equal(update_insert_trips["to_update"], True)
-    insert_mask = update_insert_trips["existing_trip_hash"].isna()
-
-    process_logger.add_metadata(
-        trip_insert_count=insert_mask.sum(),
-        trip_update_count=update_mask.sum(),
-    )
-
-    if update_mask.sum() > 0:
-        update_cols = [
-            "trip_hash",
-            "stop_count",
-            "static_trip_id_guess",
-            "static_start_time",
-            "static_stop_count",
-            "first_last_station_match",
-        ]
-        db_manager.execute_with_data(
-            sa.update(VehicleTrips.__table__).where(
-                VehicleTrips.trip_hash == sa.bindparam("b_trip_hash"),
-            ),
-            update_insert_trips.loc[update_mask, update_cols].rename(
-                columns={"trip_hash": "b_trip_hash"}
-            ),
-        )
-
-    if insert_mask.sum() > 0:
-        insert_cols = [
-            "trip_hash",
-            "direction_id",
-            "route_id",
-            "trunk_route_id",
-            "start_date",
-            "start_time",
-            "vehicle_id",
-            "stop_count",
-            "static_trip_id_guess",
-            "static_start_time",
-            "static_stop_count",
-            "first_last_station_match",
-        ]
-        db_manager.execute_with_data(
-            sa.insert(VehicleTrips.__table__),
-            update_insert_trips.loc[insert_mask, insert_cols],
-        )
-
-    process_logger.log_complete()
-
-
 def process_metrics_table(
     db_manager: DatabaseManager,
     seed_start_date: int,
@@ -828,13 +404,22 @@ def process_metrics_table(
 # pylint: enable=R0914
 
 
-def process_trips_and_metrics(db_manager: DatabaseManager) -> None:
+def process_metrics(
+    db_manager: DatabaseManager, events: pandas.DataFrame
+) -> None:
     """
-    insert and update trips and metrics tables, one day at a time
+    insert and update metrics table
     """
-    try:
-        seed_data = get_seed_data(db_manager)
-        process_tables(db_manager, seed_data)
+    for start_date in events["start_date"].unique():
+        timestamps = [
+            int(s_d)
+            for s_d in events.loc[
+                events["start_date"] == start_date, "fk_static_timestamp"
+            ].unique()
+        ]
 
-    except Exception as e:
-        logging.exception(e)
+        process_metrics_table(
+            db_manager,
+            seed_start_date=int(start_date),
+            seed_timestamps=timestamps,
+        )
