+++ python_src/src/lamp_py/performance_manager/pipeline.py
@@ -17,7 +17,6 @@ from lamp_py.runtime_utils.alembic_migration import run_alembic_migration
 
 from .l0_gtfs_static_table import process_static_tables
 from .l0_gtfs_rt_events import process_gtfs_rt_files
-from .l1_rt_metrics import process_trips_and_metrics
 
 logging.getLogger().setLevel("INFO")
 
@@ -96,7 +95,6 @@ def main(args: argparse.Namespace) -> None:
         try:
             process_static_tables(db_manager)
             process_gtfs_rt_files(db_manager)
-            process_trips_and_metrics(db_manager)
 
             process_logger.log_complete()
         except Exception as exception:
